eval/pyproject.toml
[project]
name = "spatial-ai-eval"
version = "0.1.0"
description = "Model evaluation with lm-eval"
requires-python = ">=3.11"
dependencies = [
    "lm-eval[vllm]>=0.4.7",
    "pyyaml>=6.0.3",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "llama-index>=0.12.42",
    "llama-index-embeddings-huggingface>=0.5.5",
]

[tool.uv.sources]
lm-eval = { git = "https://github.com/EleutherAI/lm-evaluation-harness.git" }
